Move effects to separate handler
