+++ src/main/java/il/co/codeguru/corewars8086/gui/WarFrame.java
@@ -1,6 +1,6 @@
 package il.co.codeguru.corewars8086.gui;
 
-import il.co.codeguru.corewars8086.gui.effects.*;
+import il.co.codeguru.corewars8086.gui.effects.WarEffectsHandler;
 import il.co.codeguru.corewars8086.memory.MemoryEventListener;
 import il.co.codeguru.corewars8086.memory.RealModeAddress;
 import il.co.codeguru.corewars8086.utils.Unsigned;
@@ -8,8 +8,6 @@ import il.co.codeguru.corewars8086.war.*;
 
 import java.awt.*;
 import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.Map;
 
 import javax.swing.*;
 
@@ -58,16 +56,12 @@ public class WarFrame extends JFrame
     
 
     private JSlider speedSlider;
-    
-    private JCheckBox enableEffectsCheckBox;
 
     private final Competition competition;
 
 	private MemoryFrame memoryFrame;
 	
-	private GameEventDetector eventDetector;
-	private Map<String, Integer> warriorNameToIndex;
-	private Map<String, Long> recentDeathEvents; // Track recently triggered death events to prevent duplicates
+	private WarEffectsHandler effectsHandler;
 
     public WarFrame(final Competition competition) {
         super("CodeGuru Extreme - Session Viewer");
@@ -90,11 +84,6 @@ public class WarFrame extends JFrame
         warCanvas = new Canvas();
         canvasPanel.add(warCanvas);
         warZone.add(canvasPanel, BorderLayout.CENTER);
-        
-        // Initialize event detector
-        eventDetector = null;
-        warriorNameToIndex = new HashMap<>();
-        recentDeathEvents = new HashMap<>();
 
         //warZone.add(new JLabel(new ImageIcon("images/warzone.jpg")), BorderLayout.NORTH);
         mainPanel.add(warZone, BorderLayout.CENTER);
@@ -124,9 +113,14 @@ public class WarFrame extends JFrame
         });
         buttonPanel.add(speedSlider);
         buttonPanel.add(Box.createHorizontalStrut(20));
-        enableEffectsCheckBox = new JCheckBox("Enable Effects", true);
+        JCheckBox enableEffectsCheckBox = new JCheckBox("Enable Effects", true);
         enableEffectsCheckBox.setForeground(Color.BLACK);
         enableEffectsCheckBox.setOpaque(false);
+        enableEffectsCheckBox.addActionListener(e -> {
+            if (effectsHandler != null) {
+                effectsHandler.setEffectsEnabled(enableEffectsCheckBox.isSelected());
+            }
+        });
         buttonPanel.add(enableEffectsCheckBox);
         nRoundNumber = 0;
         infoZone.add(buttonPanel, BorderLayout.SOUTH);
@@ -190,7 +184,10 @@ public class WarFrame extends JFrame
         getContentPane().setBackground(Color.BLACK);
         getContentPane().add(mainPanel, BorderLayout.CENTER);
         //getContentPane().add(new JLabel(new ImageIcon("images/title2.png")), BorderLayout.EAST);
-        getContentPane().add(infoZone, BorderLayout.SOUTH);		
+        getContentPane().add(infoZone, BorderLayout.SOUTH);
+        
+        // Initialize effects handler
+        effectsHandler = new WarEffectsHandler(warCanvas, competition);
     }
 
     /** Add a message to the message zone */
@@ -217,130 +214,7 @@ public class WarFrame extends JFrame
         			Unsigned.unsignedShort(ipInsideArena),
         			(byte)competition.getCurrentWarrior());
         	
-        	// Detect special events
-        	if (eventDetector != null && competition.getCurrentWarrior() >= 0) {
-        		GameEventDetector.GameEvent event = eventDetector.detectMemoryWrite(
-        			address, 
-        			competition.getCurrentWarrior()
-        		);
-        		
-        		if (event != null) {
-        			triggerEffectForEvent(event);
-        		}
-        	}
-        }
-    }
-    
-    /**
-     * Triggers visual effects based on detected game events.
-     */
-    private void triggerEffectForEvent(GameEventDetector.GameEvent event) {
-        // Check if effects are enabled
-        if (!enableEffectsCheckBox.isSelected()) {
-            return;
-        }
-        
-        int[] screenPos = GameEventDetector.arenaOffsetToScreen(
-            event.arenaOffset,
-            Canvas.BOARD_SIZE,
-            Canvas.DOT_SIZE
-        );
-        int x = screenPos[0];
-        int y = screenPos[1];
-        
-        Color warriorColor = warCanvas.getColorForWarrior(
-            event.warriorIndex >= 0 ? event.warriorIndex : 0
-        );
-        
-        int canvasWidth = Canvas.BOARD_SIZE * Canvas.DOT_SIZE;
-        int canvasHeight = Canvas.BOARD_SIZE * Canvas.DOT_SIZE;
-        
-        switch (event.type) {
-            case BOMB_USED:
-                // Explosion + particles
-                warCanvas.addVisualEffect(new ExplosionEffect(x, y, 30, Color.CYAN, 800));
-                warCanvas.addVisualEffect(new ParticleEffect(x, y, Color.ORANGE, 20, 1000));
-                break;
-                
-            case ZOMBIE_CAPTURED:
-                // Get capturing warrior's info from the event
-                String captorName = "Unknown";
-                Color captorColor = Color.YELLOW;
-                if (competition.getCurrentWar() != null && event.captorIndex >= 0) {
-                    Warrior captor = competition.getCurrentWar().getWarrior(event.captorIndex);
-                    if (captor != null) {
-                        captorName = captor.getName();
-                        captorColor = warCanvas.getColorForWarrior(event.captorIndex);
-                    }
-                }
-                
-                // BIG celebration for zombie capture!
-                // Multiple fireworks bursts
-                warCanvas.addVisualEffect(new FireworksEffect(x, y, captorColor, 5, 2000));
-                warCanvas.addVisualEffect(new FireworksEffect(x - 20, y - 15, captorColor, 4, 1800));
-                warCanvas.addVisualEffect(new FireworksEffect(x + 20, y - 15, captorColor, 4, 1800));
-                
-                // Particle explosion
-                warCanvas.addVisualEffect(new ParticleEffect(x, y, captorColor, 25, 1500));
-                
-                // Pulsing glow
-                warCanvas.addVisualEffect(new GlowEffect(x, y, 50, captorColor, 5, 1500));
-                
-                // Big bold floating text - goes all the way up
-                warCanvas.addVisualEffect(new FloatingTextEffect(
-                    "ZOMBIE CAPTURED!", x, y - 20, new Color(255, 215, 0),
-                    new Font("Arial", Font.BOLD, 28), 250, 10000,
-                    canvasWidth, canvasHeight
-                ));
-                warCanvas.addVisualEffect(new FloatingTextEffect(
-                    captorName + " controls " + event.warriorName, x, y + 10, captorColor,
-                    new Font("Arial", Font.BOLD, 18), 220, 10000,
-                    canvasWidth, canvasHeight
-                ));
-                break;
-                
-            case WARRIOR_DEATH:
-                // Prevent duplicate death events (within 2 seconds)
-                String deathKey = event.warriorName + "_death";
-                long currentTime = System.currentTimeMillis();
-                Long lastDeathTime = recentDeathEvents.get(deathKey);
-                if (lastDeathTime != null && (currentTime - lastDeathTime) < 2000) {
-                    // Skip duplicate death event
-                    return;
-                }
-                recentDeathEvents.put(deathKey, currentTime);
-                
-                // Check if dead warrior is a zombie
-                boolean isZombie = false;
-                if (competition.getCurrentWar() != null && event.warriorIndex >= 0) {
-                    Warrior deadWarrior = competition.getCurrentWar().getWarrior(event.warriorIndex);
-                    if (deadWarrior != null && deadWarrior.isZombie()) {
-                        isZombie = true;
-                    }
-                }
-                
-                // Glow blink effect
-                warCanvas.addVisualEffect(new GlowEffect(x, y, 35, warriorColor, 4, 1000));
-                
-                // Permanent grave marker
-                warCanvas.addVisualEffect(new GraveEffect(x, y, event.warriorName, warriorColor, isZombie));
-                
-                if (isZombie) {
-                    // Zombie killed - show specific message (bigger, longer, floats up)
-                    warCanvas.addVisualEffect(new FloatingTextEffect(
-                        event.warriorName + " DESTROYED!", x, y, new Color(150, 150, 150),
-                        new Font("Arial", Font.BOLD, 20), 200, 8000,
-                        canvasWidth, canvasHeight
-                    ));
-                } else {
-                    // Regular warrior death (bigger, longer, floats up)
-                    warCanvas.addVisualEffect(new FloatingTextEffect(
-                        event.message, x, y, warriorColor,
-                        new Font("Arial", Font.BOLD, 24), 250, 10000,
-                        canvasWidth, canvasHeight
-                    ));
-                }
-                break;
+        	effectsHandler.onMemoryWrite(competition.getCurrentWarrior(), address);
         }
     }
 
@@ -349,24 +223,7 @@ public class WarFrame extends JFrame
         addMessage("=== Session started ===");
         nameListModel.clear();
         warCanvas.clear();
-        warCanvas.clearEffects();
-        recentDeathEvents.clear(); // Clear recent death events for new war
-        
-        // Initialize event detector
-        if (competition.getCurrentWar() != null) {
-            eventDetector = new GameEventDetector(competition.getCurrentWar());
-            eventDetector.updateWarriorRegions();
-            warriorNameToIndex.clear();
-            
-            // Build name to index mapping
-            for (int i = 0; i < competition.getCurrentWar().getNumWarriors(); i++) {
-                Warrior warrior = competition.getCurrentWar().getWarrior(i);
-                if (warrior != null) {
-                    warriorNameToIndex.put(warrior.getName(), i);
-                }
-            }
-        }
-        
+        effectsHandler.onWarStart();
         if (competition.getCurrentWar().isPaused()){
 			btnPause.setText("Resume");
 			btnSingleRound.setEnabled(true);
@@ -378,121 +235,30 @@ public class WarFrame extends JFrame
         roundNumber.setText(Integer.toString(nRoundNumber));
         roundNumber.repaint();		
 
-        // War end effect
-        int centerX = Canvas.BOARD_SIZE * Canvas.DOT_SIZE / 2;
-        int centerY = Canvas.BOARD_SIZE * Canvas.DOT_SIZE / 2;
-        int canvasWidth = Canvas.BOARD_SIZE * Canvas.DOT_SIZE;
-        int canvasHeight = Canvas.BOARD_SIZE * Canvas.DOT_SIZE;
-        String victoryText;
-        Color winnerColor = Color.YELLOW;
-        
-        // Try to get winner's color
-        if (competition.getCurrentWar() != null && winners != null && !winners.isEmpty()) {
-            String[] winnerNames = winners.split(",");
-            if (winnerNames.length > 0) {
-                String firstWinner = winnerNames[0].trim();
-                // Find winner's index
-                for (int i = 0; i < competition.getCurrentWar().getNumWarriors(); i++) {
-                    Warrior w = competition.getCurrentWar().getWarrior(i);
-                    if (w != null && w.getName().equals(firstWinner)) {
-                        winnerColor = warCanvas.getColorForWarrior(i);
-                        break;
-                    }
-                }
-            }
-        }
-        
         switch (reason) {
             case SINGLE_WINNER:
                 addMessage(nRoundNumber,
                     "Session over: The winner is " + winners + "!");
-                victoryText = "WINNER: " + winners + "!";
-                
-                if (enableEffectsCheckBox.isSelected()) {
-                    // Dramatic winner announcement with multiple effects
-                    // Gold color for winner effects
-                    Color goldColor = new Color(255, 215, 0);
-                    
-                    // Celebration particles from center
-                    warCanvas.addVisualEffect(new ParticleEffect(
-                        centerX, centerY, goldColor, 30, 2000
-                    ));
-                    
-                    // Multiple pulse rings
-                    warCanvas.addVisualEffect(new PulseEffect(
-                        centerX, centerY, 100, goldColor, 5, 2000
-                    ));
-                    
-                    // Large prominent winner text (bigger, stays longer)
-                    warCanvas.addVisualEffect(new FloatingTextEffect(
-                        "VICTORY!", centerX, centerY - 40, goldColor,
-                        new Font("Arial", Font.BOLD, 42), 100, 12000,
-                        canvasWidth, canvasHeight
-                    ));
-                    
-                    // Winner name text (bigger, stays longer)
-                    warCanvas.addVisualEffect(new FloatingTextEffect(
-                        victoryText, centerX, centerY + 30, winnerColor,
-                        new Font("Arial", Font.BOLD, 38), 80, 12000,
-                        canvasWidth, canvasHeight
-                    ));
-                }
                 break;
-                
             case MAX_ROUND_REACHED:
                 addMessage(nRoundNumber,
                     "Maximum round reached: The winners are " + winners + "!");
-                victoryText = "Winners: " + winners;
-                
-                if (enableEffectsCheckBox.isSelected()) {
-                    // Less dramatic but still celebratory
-                    warCanvas.addVisualEffect(new ParticleEffect(
-                        centerX, centerY, new Color(150, 150, 255), 20, 1500
-                    ));
-                    
-                    warCanvas.addVisualEffect(new FloatingTextEffect(
-                        "TIME UP", centerX, centerY - 20, new Color(200, 200, 255),
-                        new Font("Arial", Font.BOLD, 36), 80, 10000,
-                        canvasWidth, canvasHeight
-                    ));
-                    
-                    warCanvas.addVisualEffect(new FloatingTextEffect(
-                        victoryText, centerX, centerY + 30, winnerColor,
-                        new Font("Arial", Font.BOLD, 26), 60, 10000,
-                        canvasWidth, canvasHeight
-                    ));
-                }
                 break;
-                
             case ABORTED:
                 addMessage(nRoundNumber,
                     "Session aborted: The winners are " + winners + "!");
-                victoryText = "Aborted";
                 break;
             default:
                 throw new RuntimeException();			
         }
+        
+        effectsHandler.onWarEnd(reason, winners, nRoundNumber);
     }	
 
     /** @see CompetitionEventListener#onRound(int) */
     public void onRound(int round) {
         nRoundNumber = round;
-        
-        // Update event detector with current round
-        // This enables bomb detection after a grace period to avoid false positives
-        // from rapid round execution in early game
-        if (eventDetector != null) {
-            eventDetector.onRound(round);
-            
-            // Check for zombie capture (zombie executing code written by another warrior)
-            if (enableEffectsCheckBox.isSelected()) {
-                GameEventDetector.GameEvent zombieEvent = eventDetector.checkZombieCapture();
-                if (zombieEvent != null) {
-                    triggerEffectForEvent(zombieEvent);
-                }
-            }
-        }
-        
+        effectsHandler.onRound(round);
         if ((nRoundNumber % 1000) == 0) {
             roundNumber.setText(Integer.toString(nRoundNumber));
             roundNumber.repaint();
@@ -505,11 +271,7 @@ public class WarFrame extends JFrame
     public void onWarriorBirth(String warriorName) {
         addMessage(nRoundNumber, warriorName + " enters the arena.");
         nameListModel.addElement(new WarriorInfo(warriorName));
-        
-        // Update event detector
-        if (eventDetector != null) {
-            eventDetector.updateWarriorRegions();
-        }
+        effectsHandler.onWarriorBirth();
     }
 
     /** @see CompetitionEventListener#onWarriorDeath(String) */
@@ -531,28 +293,7 @@ public class WarFrame extends JFrame
             }
         });
         
-        // Trigger death effect only once - check if warrior is already dead to avoid duplicates
-        if (eventDetector != null && competition.getCurrentWar() != null) {
-            Integer warriorIndex = warriorNameToIndex.get(warriorName);
-            if (warriorIndex == null) {
-                // Find index by name
-                for (int i = 0; i < competition.getCurrentWar().getNumWarriors(); i++) {
-                    Warrior w = competition.getCurrentWar().getWarrior(i);
-                    if (w != null && w.getName().equals(warriorName)) {
-                        warriorIndex = i;
-                        warriorNameToIndex.put(warriorName, i);
-                        break;
-                    }
-                }
-            }
-            
-            if (warriorIndex != null) {
-                GameEventDetector.GameEvent event = eventDetector.createDeathEvent(
-                    warriorName, warriorIndex
-                );
-                triggerEffectForEvent(event);
-            }
-        }
+        effectsHandler.onWarriorDeath(warriorName);
     }	
 
     /**
@@ -628,12 +369,7 @@ public class WarFrame extends JFrame
 	@Override
 	public void onEndRound() {
 		this.warCanvas.deletePointers();
-		
-		// Update event detector warrior regions in case they've changed
-		if (eventDetector != null && competition.getCurrentWar() != null) {
-			eventDetector.updateWarriorRegions();
-		}
-		
+		effectsHandler.onEndRound();
 		for (int i = 0; i < this.competition.getCurrentWar().getNumWarriors(); i++)
 			if (this.competition.getCurrentWar().getWarrior(i).isAlive()) {
 				short ip = this.competition.getCurrentWar().getWarrior(i).getCpuState().getIP();
